v0.5.0 - Major overhaul with new Url Class
The new Url class puts all 4chan API URLs in one handy class. This way, URL construction is done in this class and can be easily reused. In fact, it might be helpful in our BASC-Archiver.

This replaces the former URL dictionary created by Danneh.

A Url Class also makes it possible for derived classes to inherit and override the typical 4chan API URL settings, so 8chan/vichan or 420chan can be supported by just tweaking the URLs a bit.